MacroVision/backend/app/services/prediction.py
import torch
from PIL import Image
from io import BytesIO
from torchvision import transforms
from app.models.classifier import FoodClassifier
from ..schemas import PredictionResult

# Initialize models (simplified for example)
food_model = None
portion_model = None

def load_models():
    """Load trained models into memory"""
    global food_model, portion_model
    
    # Initialize with your actual number of food classes
    food_model = FoodClassifier(num_classes=10)
    # Load your trained weights
    food_model.load_state_dict(torch.load("/app/models/food_classifier.pth"))
    food_model.eval()
    
    # Similarly for portion estimator (implementation depends on your model)

async def predict_macros(image: Image) -> PredictionResult:
    """Predict nutrition macros from food image"""
    if food_model is None:
        load_models()
    
    # Preprocess image
    transform = transforms.Compose([
        transforms.Resize(256),
        transforms.CenterCrop(224),
        transforms.ToTensor(),
        transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225]),
    ])
    image_tensor = transform(image).unsqueeze(0)
    
    # Predict (example implementation)
    with torch.no_grad():
        food_class = torch.argmax(food_model(image_tensor)).item()
    
    # Placeholder values - replace with your actual implementation
    return PredictionResult(
        food="example_food",
        portion_g=100.0,
        macros={"protein": 20.0, "carbs": 30.0, "fat": 10.0}
    )