MacroVision/backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
redis==4.5.5
numpy==1.24.4  # Pinned to avoid NumPy 2.0 compatibility issues